LOC-6: Use %v for temp variables
Need to use something like %v_wt, instead of plain ?wt, so that
we get the temp variable number included in the names, and will
not be suing the same temp variable in multiple clauses. That
would go wrong with searches like
  @and
    @atter 1=bf.author a
    @atter 1=bf.author b
With this patch, this searches for works that have an author that
has an 'a', and an author that has a 'b'. Without, it would be
searching for a work that has an author with both 'a' and 'b'.

Unfortunately this makes the config file even less readable.